analytics-engine/requirements.txt
kafka-python==2.0.2
pandas==2.2.0
streamlit==1.37.0
plotly==5.18.0
scikit-learn==1.4.0
xgboost==2.0.3
tensorflow==2.15.0
websocket-client==1.7.0
